+++ examples/blog/routes/index.js
@@ -27,13 +27,34 @@ Router.get('/admin', (req, res) => {
   if (session.has('superuser')) {
     res.redirect('/admin/dashboard');
   }
-  
   res.render('admin/index', {title: 'Blog'});
 });
 
 // Admin Sign-in
 Router.post('/admin/signin', (req, res) => {
-  res.end('oke');
+  if (session.has('superuser')) {
+    res.redirect('/admin/dashboard');
+  } else {
+    if (req.body.username == 'admin' && req.body.password == 123) {
+      session.set('superuser', req.body.username);
+      res.redirect('/admin/dashboard');
+    } else {
+      res.end('login gagal!');
+    }
+  }
+});
+
+// Dashboard
+Router.get('/admin/dashboard', (req, res) => {
+  if (!session.has('superuser')) {
+    res.redirect('/admin');
+  }
+  res.render('admin/dashboard', {
+    title: 'Blog',
+    data: [{
+      name: session.get('superuser')
+    }]
+  });
 });
 
 module.exports = Router;